pyproject.toml
[project]
name = "prefect-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.128.0",
    "fastmcp>=2.14.4",
    "litellm>=1.81.4",
    "mcp>=1.26.0",
    "numpy>=2.4.1",
    "prefect>=3.6.13",
    "python-dotenv>=1.2.1",
    "scipy>=1.17.0",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
